requirements.txt
alembic==1.11.1
annotated-types==0.7.0
anyio==3.6.2
bcrypt==4.2.0
certifi==2022.12.7
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.115.4
fastapi-login==1.10.2
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.6
MarkupSafe==2.1.2
orjson==3.8.10
passlib==1.7.4
psycopg2-binary==2.9.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.11
starlette==0.41.2
typing_extensions==4.12.2
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0